requirements.txt
numpy
pandas
torch
scikit-learn
alpaca-py
ta
python-dotenv
tqdm
